requirements.txt
anyio==4.2.0
asgiref==3.7.2
backoff==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.1
django-autocomplete-light==3.9.7
django-bootstrap-input-group==1.0.1
django-bootstrap5==23.4
django-environ==0.11.2
django-polymorphic==3.1.0
django-simple-history==3.4.0
django-storages==1.14.2
exceptiongroup==1.2.0
gql==3.5.0
graphql-core==3.2.3
gunicorn==21.2.0
idna==3.6
multidict==6.0.5
mysqlclient==2.2.3
packaging==23.2
pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4
